+++ tests/e2e/game/game-won-by-abominable-sectarian.test.js
@@ -1,0 +1,91 @@
+const { describe, it, before, after } = require("mocha");
+const chai = require("chai");
+const chaiHttp = require("chai-http");
+const app = require("../../../app");
+const Config = require("../../../config");
+const { resetDatabase } = require("../../../src/helpers/functions/Test");
+
+chai.use(chaiHttp);
+const { expect } = chai;
+
+const credentials = { email: "test@test.fr", password: "secret" };
+const originalPlayers = [
+    { name: "Dag", role: "abominable-sectarian", group: "1" },
+    { name: "Dig", role: "werewolf", group: "1" },
+    { name: "Deg", role: "villager", group: "2" },
+    { name: "Dog", role: "villager", group: "2" },
+];
+let server, token, game, players;
+
+describe("α - Tiny game of 4 players in which the abominable sectarian wins because his group survived", () => {
+    before(done => resetDatabase(done));
+    before(done => {
+        server = app.listen(3000, done);
+    });
+    after(done => resetDatabase(done));
+    it("👤 Creates new user (POST /users)", done => {
+        chai.request(server)
+            .post("/users")
+            .auth(Config.app.basicAuth.username, Config.app.basicAuth.password)
+            .send(credentials)
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                done();
+            });
+    });
+    it("🔑 Logs in successfully (POST /users/login)", done => {
+        chai.request(server)
+            .post(`/users/login`)
+            .auth(Config.app.basicAuth.username, Config.app.basicAuth.password)
+            .send(credentials)
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                token = res.body.token;
+                done();
+            });
+    });
+    it("🎲 Creates game with JWT auth (POST /games)", done => {
+        chai.request(server)
+            .post("/games")
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ players: originalPlayers, options: { roles: { sheriff: { isEnabled: false } } } })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                done();
+            });
+    });
+    it("🐺 Werewolf eats the first villager (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(server)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "werewolves", action: "eat", targets: [{ player: players[2]._id }] })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                done();
+            });
+    });
+    it("👪 All vote for the second villager (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(server)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "all", action: "vote", votes: [{ from: players[0]._id, for: players[3]._id }] })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                expect(game.players[3].isAlive).to.be.false;
+                expect(game.players[3].murdered).to.deep.equals({ by: "all", of: "vote" });
+                done();
+            });
+    });
+    it("🎲 Game is WON by the abominable sectarian because the other group is dead", done => {
+        expect(game.status).to.equal("done");
+        expect(game.won.by).to.equal("abominable-sectarian");
+        expect(game.won.players).to.be.an("array").lengthOf(1);
+        expect(game.won.players[0]._id).to.equal(players[0]._id);
+        done();
+    });
+});